+++ src/ScottPlot/Palettes/DarkPastel.cs
@@ -1,0 +1,15 @@
+﻿/* This a qualitative 8-color palette generated using https://colorbrewer2.org
+ * © Cynthia Brewer, Mark Harrower and The Pennsylvania State University
+ * This palette is the lighter-color version of the 'Dark' palette.
+ */
+namespace ScottPlot.Drawing.Colorsets
+{
+    class DarkPastel : HexColorset, IPalette
+    {
+        public override string[] hexColors => new string[]
+        {
+            "#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854",
+            "#ffd92f", "#e5c494", "#b3b3b3",
+        };
+    }
+}