+++ dev/changelog.md
@@ -5,7 +5,7 @@
 * Style: Refactored to use static classes instead of enumeration members (#1299, #1291)
 * NuGet: Improved System.Drawing.Common dependencies in user control packages (#1311, #1310) _Thanks @Kritner_
 * New Palettes:
-  * `DarkLight` 8-color palette of LCD and print-friendly colors (#1313) _Thanks @arthurits_
+  * `DarkPastel` 8-color palette of LCD and print-friendly colors (#1313) _Thanks @arthurits_
 
 ## ScottPlot 4.1.23
 * NuGet: use deterministic builds, add source link support, and include compiler flags (#1285)